Validation code for AMROC rocket
This code is the AMROC version of the other c code for validation of the code. In this case, I am using American SI units as using other systems resulting in wrong values. So to validate the formula I have written this code with that in mind. Currently, all values check out but I need to confirm the mass flow rate is calculated correctly as I am currently using a known value of mass flow rate calculated by taking the weight of fuel and dividing it by the burn time. After ensuring that Mdot is calculated correctly I can fully validate the code :)