Parameters moved to struc + bug in height?
in preparation for a combination of codes, I have moved all rocket parameters to a struct for easier reading. I have also discovered a bug where the height of the rocket will not decrease below 10000 and I have found a fix to make the dynamics conditional on both T >0 and a turnover height eg 15000 seems to work well currently but this needs to be investigated further.