Re-formatted code to run states instead of vectors
Re-formatted code to run states instead of vectors. This means supporting code is super reduced instead of redundant repeated lines. This also makes more sense for tracking values for one instance. I still have problems with Velocity and height not appearing as expected. there seems to be no gravity effect once again it just continues up infinitely but the start of the height plot looks pretty good. Quaternions are re-evaluated as well and seem to be working well.